Moved NBitcoin's JsonConverters to Stratis.Bitcoin
