+++ src/NBitcoin.Tests/util_tests.cs
@@ -7,7 +7,6 @@ using System.Text;
 using NBitcoin.BouncyCastle.Math;
 using NBitcoin.Crypto;
 using NBitcoin.DataEncoders;
-using NBitcoin.JsonConverters;
 using NBitcoin.OpenAsset;
 using Newtonsoft.Json.Linq;
 using Xunit;
@@ -664,8 +663,9 @@ namespace NBitcoin.Tests
             result = Network.Parse(address.Base58, null);
             Assert.True(result.Network == Network.TestNet);
 
-            string str = Serializer.ToString(new DummyClass() { ExtPubKey = new ExtKey().Neuter().GetWif(Network.RegTest) }, Network.RegTest);
-            Assert.NotNull(Serializer.ToObject<DummyClass>(str, Network.RegTest));
+            // TODO Move to Stratis.Bitcoin
+            //string str = Serializer.ToString(new DummyClass() { ExtPubKey = new ExtKey().Neuter().GetWif(Network.RegTest) }, Network.RegTest);
+            //Assert.NotNull(Serializer.ToObject<DummyClass>(str, Network.RegTest));
         }
 
         [Fact]
