+++ src/Stratis.Bitcoin/Utilities/JsonConverters/CoinJsonConverter.cs
@@ -1,9 +1,10 @@
 ﻿using System;
 using System.Reflection;
+using NBitcoin;
 using NBitcoin.OpenAsset;
 using Newtonsoft.Json;
 
-namespace NBitcoin.JsonConverters
+namespace Stratis.Bitcoin.Utilities.JsonConverters
 {
     public class CoinJsonConverter : JsonConverter
     {
