+++ src/Stratis.Bitcoin/Utilities/JsonConverters/JsonObjectException.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Newtonsoft.Json;
 
-namespace NBitcoin.JsonConverters
+namespace Stratis.Bitcoin.Utilities.JsonConverters
 {
     public class JsonObjectException : Exception
     {
