+++ src/Stratis.Bitcoin/Utilities/JsonConverters/Serializer.cs
@@ -1,7 +1,8 @@
-﻿using Newtonsoft.Json;
+﻿using NBitcoin;
+using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 
-namespace NBitcoin.JsonConverters
+namespace Stratis.Bitcoin.Utilities.JsonConverters
 {
     public class Serializer
     {
