+++ src/Stratis.Bitcoin/Utilities/JsonConverters/SignatureJsonConverter.cs
@@ -1,9 +1,10 @@
 ﻿using System;
+using NBitcoin;
 using NBitcoin.Crypto;
 using NBitcoin.DataEncoders;
 using Newtonsoft.Json;
 
-namespace NBitcoin.JsonConverters
+namespace Stratis.Bitcoin.Utilities.JsonConverters
 {
     public class SignatureJsonConverter : JsonConverter
     {
