+++ src/Stratis.Bitcoin/Utilities/JsonConverters/TxDestinationJsonConverter.cs
@@ -1,8 +1,9 @@
 ﻿using System;
+using NBitcoin;
 using NBitcoin.DataEncoders;
 using Newtonsoft.Json;
 
-namespace NBitcoin.JsonConverters
+namespace Stratis.Bitcoin.Utilities.JsonConverters
 {
     public class TxDestinationJsonConverter : JsonConverter
     {
